Creation of useSearch() & hooks folder
